formatted

+++ src/SwapTokensApp.sol
@@ -16,24 +16,24 @@ contract SwapTokensApp {
     //Variables
     address public RouterV2Address;
     //Events
+
     event SwapTokens(address tokenIn, address tokenOut, uint256 amountIn, uint256 amountOut);
-  
+
     //Constructor
 
     constructor(address RouterV2Address_) {
         RouterV2Address = RouterV2Address_;
     }
 
-    function swapTokens(uint256 amountIn_,uint256 amountOutMin_,address[] memory path_,uint256 deadline_) external{
-        
+    function swapTokens(uint256 amountIn_, uint256 amountOutMin_, address[] memory path_, uint256 deadline_) external {
         //Get first Token
         IERC20(path_[0]).safeTransferFrom(msg.sender, address(this), amountIn_);
         //Approve
         IERC20(path_[0]).approve(RouterV2Address, amountIn_);
         //Swap
-        uint[] memory amountsOut_ = IRouterV2(RouterV2Address).swapExactTokensForTokens(amountIn_, amountOutMin_, path_, msg.sender, deadline_);
+        uint256[] memory amountsOut_ =
+            IRouterV2(RouterV2Address).swapExactTokensForTokens(amountIn_, amountOutMin_, path_, msg.sender, deadline_);
 
-        emit SwapTokens(path_[0], path_[path_.length-1], amountIn_, amountsOut_[amountsOut_.length-1]);
-
+        emit SwapTokens(path_[0], path_[path_.length - 1], amountIn_, amountsOut_[amountsOut_.length - 1]);
     }
 }
